requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.1
idna==3.4
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
requests==2.31.0
urllib3==2.0.7
python-dotenv==1.0.0
firebase-admin